Refactor Room Model
The sub schemas and helper functions were moved to separate files. The
dependency paths were updated.
