+++ src/controllers/room/join.ts
@@ -1,5 +1,5 @@
 import { ApiError } from "middleware/errors";
-import makeRoomModel, { RoomId } from "models/Room";
+import makeRoomModel, { RoomId } from "models/Room/index.model";
 import DatabaseClients from "services/db";
 import makeAsyncController from "utils/reqRes/asyncController";
 // import APP_ENV_VARS from "utils/setup/env";
